requirements.txt
python-dotenv  
pyautogen[autobuild]
